+++ src/client/keys.h
@@ -22,37 +22,47 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 #include "keycodes.h"
 
-typedef struct {
-	qboolean	down;
-	int			repeats;		// if > 1, it is autorepeating
-	char		*binding;
-} qkey_t;
+typedef struct
+{
+  qbool down;
+  qbool bound;
+  qint repeats; //if > 1, it is autorepeating
+  qchar *binding;
+}
+qkey_t;
 
-extern	qboolean	key_overstrikeMode;
-extern	qkey_t		keys[MAX_KEYS];
+extern qbool key_overstrikeMode;
+extern qkey_t keys[MAX_KEYS];
 
-// NOTE TTimo the declaration of field_t and Field_Clear is now in qcommon/qcommon.h
-void Field_KeyDownEvent( field_t *edit, int key );
-void Field_CharEvent( field_t *edit, int ch );
-void Field_Draw( field_t *edit, int x, int y, int width, qboolean showCursor, qboolean noColorEscape );
-void Field_BigDraw( field_t *edit, int x, int y, int width, qboolean showCursor, qboolean noColorEscape );
+extern qint anykeydown;
 
-#define		COMMAND_HISTORY		32
-extern	field_t	historyEditLines[COMMAND_HISTORY];
+//NOTE TTimo the declaration of field_t and Field_Clear is now in qcommon/qcommon.h
 
-extern	field_t	g_consoleField;
-extern	field_t	chatField;
-extern	int				anykeydown;
-extern	qboolean	chat_team;
-extern	int			chat_playerNum;
+void
+Key_WriteBindings(fileHandle_t f);
+void
+Key_SetBinding(qint keynum, const qchar *binding);
+const char *
+Key_GetBinding(qint keynum);
+void
+Key_ParseBinding(qint key, qbool down, unsigned time);
 
-void Key_WriteBindings( fileHandle_t f );
-void Key_SetBinding( int keynum, const char *binding );
-char *Key_GetBinding( int keynum );
-qboolean Key_IsDown( int keynum );
-qboolean Key_GetOverstrikeMode( void );
-void Key_SetOverstrikeMode( qboolean state );
-void Key_ClearStates( void );
-int Key_GetKey(const char *binding);
-void Key_KeynumToStringBuf( int keynum, char *buf, int buflen );
-void Key_GetBindingBuf( int keynum, char *buf, int buflen );
+qint
+Key_GetKey(const qchar *binding);
+const qchar *
+Key_KeynumToString(qint keynum);
+qint
+Key_StringToKeynum(const qchar *str);
+
+qbool
+Key_IsDown(qint keynum);
+void
+Key_ClearStates(void);
+
+qbool
+Key_GetOverstrikeMode(void);
+void
+Key_SetOverstrikeMode(qbool state);
+
+void
+Com_InitKeyCommands(void);